+++ packages/adapter/src/lib/adapter/adapter.ts
@@ -2765,14 +2765,22 @@ export class AdapterClass extends EventEmitter {
         this._intervals.delete(interval as NodeJS.Timeout);
     }
 
-    setObject(id: string, obj: ioBroker.SettableObject, callback?: ioBroker.SetObjectCallback): Promise<void>;
+    setObject(
+        id: string,
+        obj: ioBroker.SettableObject,
+        callback?: ioBroker.SetObjectCallback,
+    ): Promise<ioBroker.NonNullCallbackReturnTypeOf<ioBroker.SetObjectCallback> | void>;
     setObject(
         id: string,
         obj: ioBroker.SettableObject,
         options: unknown,
         callback?: ioBroker.SetObjectCallback,
-    ): Promise<void>;
-    setObject(id: string, obj: ioBroker.SettableObject, callback?: ioBroker.SetObjectCallback): Promise<void>;
+    ): Promise<ioBroker.NonNullCallbackReturnTypeOf<ioBroker.SetObjectCallback> | void>;
+    setObject(
+        id: string,
+        obj: ioBroker.SettableObject,
+        callback?: ioBroker.SetObjectCallback,
+    ): Promise<ioBroker.NonNullCallbackReturnTypeOf<ioBroker.SetObjectCallback> | void>;
     /**
      * Creates or overwrites an object in objectDB.
      *
@@ -2803,7 +2811,12 @@ export class AdapterClass extends EventEmitter {
      *            }
      *        ```
      */
-    setObject(id: unknown, obj: unknown, options: unknown, callback?: unknown): Promise<void> | void {
+    setObject(
+        id: unknown,
+        obj: unknown,
+        options: unknown,
+        callback?: unknown,
+    ): Promise<ioBroker.NonNullCallbackReturnTypeOf<ioBroker.SetObjectCallback> | void> {
         if (typeof options === 'function') {
             callback = options;
             options = null;
@@ -2819,7 +2832,9 @@ export class AdapterClass extends EventEmitter {
         return this._setObject({ id, obj: obj as ioBroker.SettableObject, options, callback });
     }
 
-    private async _setObject(options: InternalSetObjectOptions): Promise<void> {
+    private async _setObject(
+        options: InternalSetObjectOptions,
+    ): Promise<ioBroker.NonNullCallbackReturnTypeOf<ioBroker.SetObjectCallback> | void> {
         if (!this._defaultObjs) {
             this._defaultObjs = (await import('./defaultObjs.js')).createDefaults();
         }
@@ -2844,6 +2859,7 @@ export class AdapterClass extends EventEmitter {
                 this._utils.validateId(options.id, false, null);
             } catch (e) {
                 this._logger.error(tools.appendStackTrace(`${this.namespaceLog} ${e.message}`));
+                // Error is logged and silently ignored to not break older adapters
                 return;
             }
         }
@@ -3357,7 +3373,12 @@ export class AdapterClass extends EventEmitter {
      *            }
      *        ```
      */
-    setForeignObject(id: unknown, obj: unknown, options: unknown, callback?: unknown): MaybePromise {
+    setForeignObject(
+        id: unknown,
+        obj: unknown,
+        options: unknown,
+        callback?: unknown,
+    ): Promise<ioBroker.NonNullCallbackReturnTypeOf<ioBroker.SetObjectCallback> | void> | void {
         if (typeof options === 'function') {
             callback = options;
             options = null;
@@ -3386,7 +3407,9 @@ export class AdapterClass extends EventEmitter {
         return this._setForeignObject({ id, obj: obj as ioBroker.SettableObject, options, callback });
     }
 
-    private _setForeignObject(_options: InternalSetObjectOptions): MaybePromise {
+    private _setForeignObject(
+        _options: InternalSetObjectOptions,
+    ): Promise<ioBroker.NonNullCallbackReturnTypeOf<ioBroker.SetObjectCallback> | void> | void {
         const { options, callback, obj } = _options;
         let { id } = _options;
 
@@ -9316,7 +9339,7 @@ export class AdapterClass extends EventEmitter {
                 return tools.maybeCallbackWithError(callback, e);
             }
         }
-        this.#states.delState(id, callback);
+        await this.#states.delState(id, callback);
     }
 
     // external signature
@@ -9784,7 +9807,7 @@ export class AdapterClass extends EventEmitter {
 
                 subs[pattern][this.namespace]++;
                 this.outputCount++;
-                this.#states.setState(`system.adapter.${autoSubEntry}.subscribes`, JSON.stringify(subs));
+                await this.#states.setState(`system.adapter.${autoSubEntry}.subscribes`, JSON.stringify(subs));
             }
         }
 
@@ -10013,7 +10036,7 @@ export class AdapterClass extends EventEmitter {
                         delete subs[pattern];
                     }
                     this.outputCount++;
-                    this.#states.setState(`system.adapter.${autoSub}.subscribes`, JSON.stringify(subs));
+                    await this.#states.setState(`system.adapter.${autoSub}.subscribes`, JSON.stringify(subs));
                 }
             }
         }